ipv6: Switch to using new offload infrastructure
Patchwork-id: 56244
O-Subject: [RHEL 6.5 PATCHv2 BZ 801856 06/15] ipv6: Switch to using new offload infrastructure.
Bugzilla: 801856
RH-Acked-by: Thomas Graf <tgraf@redhat.com>
RH-Acked-by: Amerigo Wang <amwang@redhat.com>
RH-Acked-by: Daniel Borkmann <dborkman@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: David S. Miller <davem@redhat.com>

Upstream: 3336288a9feaa809839adbaf05778dc2f16665dc

	ipv6: Switch to using new offload infrastructure.

	Switch IPv6 protocol to using the new GRO/GSO calls and data.

	Signed-off-by: Vlad Yasevich <vyasevic@redhat.com>
	Signed-off-by: David S. Miller <davem@davemloft.net>

Conflicts:
	net/ipv6/af_inet6.c

Fixed a glaring bug in ipv6_gro_receive.  The code tried to locate
a new protocol, but never tried to fetch the options for that new
protocol.

kABI:
Same problem as with IPv4.  IPv6 isn't really white listed, but for
the sake of consistency and if it ever does become whitelisted, we
should do the same thing.
Had to change the logic in ipv6_gso_pull_exthdrs() a bit to allow for
for retries with inet_protos.

Signed-off-by: Vlad Yasevich <vyasevic@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>