mlx4: Remove FCS bytes from packet length
Patchwork-id: 56279
O-Subject: [Patch RHEL6] mlx4_en: Remove FCS bytes from packet length.
Bugzilla: 893707
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

From: Yevgeny Petrilin <yevgenyp@mellanox.co.il>

This patch was missed for rhel6.3.  I know it's late for 6.4, but the patch is:
A) upstream for a long time
B) very simple
C) makes a huge difference in performance

Without this patch, the performance of mlx4 hardware in Ethernet mode is
seriously compromised.  The problem is that without this patch, the
Ethernet byte count on packets is wrong causing the core GRO code to
refuse to work with the packets, causing the Ethernet interface to
never use GRO.  The net result is that without this patch, the maximum
Ethernet speed on the test system Mellanox used was 16GBit/s over a
40GBit/s link.  With it, maximum speed jumps to 33GBit/s throughput
on that same 40GBit/s link.

The business justification is that we have worked very hard to get
Mellanox to encourage customers to use our in-box mlx4 drivers.  With
this sort of performance problem, it is entirely understandable that
they would continue to tell their customers to use their mlx4 driver
instead of our in-box drivers.

I don't currently have a bug for this patch (I had one that I thought
was relevant, but turned out to be a different issue).  I'll create a
new bug for this issue, chase down the appropriate acks, and respond
to this thread with the bug ID.  Please review/ack/nak as appropriate.

------- Start of git commit message -------

When HW doesn't remove FCS bytes they are reported in the completion
byte count, we don't need to take them to skb.

Signed-off-by: Yevgeny Petrilin <yevgenyp@mellanox.co.il>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 4a5f4dd8595a3d3cdf75db7247b644ae37f5d460)

Signed-off-by: Doug Ledford <dledford@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>