dm-raid: Fix RAID10's check for sufficient redundancy
Patchwork-id: 56311
O-Subject: [RHEL 6.4 PATCH] DM-RAID: Fix RAID10's check for sufficient redundancy
Bugzilla: 889358
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Bug 889358:
 https://bugzilla.redhat.com/show_bug.cgi?id=889358

Brew Build:
 http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5257983

Testing:
Testing shows that without this patch it is not possible to activate a
RAID10 LV that has more than one device failure, even if they are in
different stripes.  When the patch is included, the LV can be activated
as long as the failures are in separate stripes (i.e. independent mirror
sets).

Upstream status:
 Neil Brown has pulled the patch into his tree and we are discussing
whether to backport to 3.7 also.
(http://marc.info/?l=linux-raid&m=135779784231505&w=2)  Please note that
the version number for upstream has been incremented from 1.4.1 to
1.4.2, while the version number associated for this bug fix in RHEL has
been incremented from 1.3.1 to 1.3.2.  (1.3.2 is an unused version
number upstream and should be safe for use.)  This happened because two
other features/changes have been introduced upstream that are not yet
present or planned for RHEL.  If those features are pulled into RHEL at
a later date, then the version number can be safely changed to 1.4.2.

Patch:
DM RAID:  Fix RAID10's check for sufficient redundancy

Before attempting to activate a RAID array, it is checked for sufficient
redundancy.  That is, we make sure that there are not too many failed
devices - or devices specified for rebuild - to undermine our ability to
activate the array.  The current code performs this check twice - once to
ensure there were not too many devices specified for rebuild by the user
('validate_rebuild_devices') and again after possibly experiencing a failure
to read the superblock ('analyse_superblocks').  Neither of these checks are
sufficient.  The first check is done properly but with insufficient
information about the possible failure state of the devices to make a good
determination if the array can be activated.  The second check is simply
done wrong in the case of RAID10 because it doesn't account for the
independence of the stripes (i.e. mirror sets).  The solution is to use the
properly written check ('validate_rebuild_devices'), but perform the check
after the superblocks have been read and we know which devices have failed.
This gives us one check instead of two and performs it in a location where
it can be done right.

Only RAID10 was affected and it was affected in the following ways:
- the code did not properly catch the condition where a user specified
  a device for rebuild that already had a failed device in the same mirror
  set.  (This condition would, however, be caught at a deeper level in MD.)
- the code triggers a false positive and denies activation when devices in
  independent mirror sets have failed - counting the failures as though they
  were all in the same set.

The most likely place this error was introduced (or this patch should have
been included) is in commit 4ec1e369 - first introduced in v3.7-rc1.

Signed-off-by: Jonathan Brassow <jbrassow@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>