nfsd: allow numeric idmapping
Patchwork-id: 56551
O-Subject: [Patch RHEL6.5 1/2] nfsd4: allow numeric idmapping
Bugzilla: 815908
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: "J. Bruce Fields" <bfields@redhat.com>

Mimic the client side by providing a module parameter that turns off
idmapping in the auth_sys case, for backwards compatibility with NFSv2
and NFSv3.

Unlike in the client case, we don't have any way to negotiate, since the
client can return an error to us if it doesn't like the id that we
return to it in (for example) a getattr call.

However, it has always been possible for servers to return numeric id's,
and as far as we're aware clients have always been able to handle them.

Also, in the auth_sys case clients already need to have numeric id's the
same between client and server.

Therefore we believe it's safe to default this to on; but the module
parameter is available to return to previous behavior if this proves to
be a problem in some unexpected setup.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Upstream: e9541ce8efc22c233a045f091c2b969923709038
Bugzilla: 758616
Signed-off-by: Jarod Wilson <jarod@redhat.com>