nfsd: default server numeric idmapping to off
Patchwork-id: 56552
O-Subject: [Patch RHEL6.5 2/2] nfsd4: default server numeric idmapping to off
Bugzilla: 815908
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: "J. Bruce Fields" <bfields@redhat.com>

This is an RHEL-specific patch.  Upstream the nfs server returns numeric
id's by default, but this has proved to be a problem for some clients.
Those clients are being fixed.  We want to be conservative about this in
RHEL, but we also have people with diskless setups who need numeric
idmapping--so the solution is to backport the feature but turn it off by
default.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>