scsi_debug: Fix off-by-one bug when unmapping region
Patchwork-id: 56569
O-Subject: [RHEL6.5 PATCH] [SCSI] scsi_debug: Fix off-by-one bug when unmapping region
Bugzilla: 902973
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

A backport of upstream commit bc977749e967daa56de1922cf4cb38525631c51c.
The code differs because upstream has an opening brace after the "if"
statement (due to other code), and an unnecessary line break was removed.

Bugzilla: 902973
Upstream Status: In upstream linux-stable
Brew: 5306316
Tested: Ran unmap portion of "discard.pl" test from bz880487 repeatedly

Note that RHEL6 does not have the LBPRZ mechanism mentioned below.

Upstream change description:

Currently it is possible to unmap one more block than user requested to
due to the off-by-one error in unmap_region(). This is probably due to
the fact that the end variable despite its name actually points to the
last block to unmap + 1. However in the condition it is handled as the
last block of the region to unmap.

The bug was not previously spotted probably due to the fact that the
region was not zeroed, which has changed with commit
be1dd78de5686c062bb3103f9e86d444a10ed783. With that commit we were able
to corrupt the ext4 file system on 256M scsi_debug device with LBPRZ
enabled using fstrim.

Since the 'end' semantic is the same in several functions there this
commit just fixes the condition to use the 'end' variable correctly in
that context.

Reported-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Lukas Czerner <lczerner@redhat.com>
Reviewed-by: Martin K. Petersen <martin.petersen@oracle.com>
Acked-by: Douglas Gilbert <dgilbert@interlog.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: James Bottomley <JBottomley@Parallels.com>

Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>