cputime: Avoid multiplication overflow on utime scaling
Patchwork-id: 56679
O-Subject: [RHEL6 PATCH] cputime: Avoid multiplication overflow on utime scaling
Bugzilla: 862758
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>

BZ#862758

Description (from upstream commit):
We scale stime, utime values based on rtime (sum_exec_runtime
converted to jiffies). During scaling we multiple rtime * utime,
which seems to be fine, since both values are converted to u64,
but it's not.

Let assume HZ is 1000 - 1ms tick. Process consist of 64 threads,
run for 1 day, threads utilize 100% cpu on user space. Machine
has 64 cpus.

Process rtime = utime will be 64 * 24 * 60 * 60 * 1000 jiffies,
which is 0x149970000. Multiplication rtime * utime result is
0x1a855771100000000, which can not be covered in 64 bits.

Result of overflow is stall of utime values visible in user
space (prev_utime in kernel), even if application still consume
lot of CPU time.

A solution to solve this is to perform the multiplication on
stime instead of utime. It's easy to grow the utime value fast
with a CPU bound thread in userspace for example. Now we assume
that doing so with stime is much harder. In most cases a task
shouldn't ever spend much time in kernel space as it tends to
sleep waiting for jobs completion when they take long to
achieve. IO is the typical example of that.

Hence scaling the cputime by performing the multiplication on
stime instead of utime should considerably reduce the chances of
an overflow on most workloads.

Upstream:
Queued to -tip tree
http://git.kernel.org/?p=linux/kernel/git/tip/tip.git;a=commit;h=62188451f0d63add7ad0cd2a1ae269d600c1663d

Brew:
https://brewweb.devel.redhat.com/taskinfo?taskID=5345314

Testing:
First version of the patch was tested by me and bug reporter. It fixes
the problem. This patch differ from original, but the principle of the
fix is the same. I'm certain it fixes the problem, so did only smoke
testing of it.

Signed-off-by: Jarod Wilson <jarod@redhat.com>