audit: __link_path_walk should call audit_inode with original "name"
Patchwork-id: 56928
O-Subject: [RHEL6.5 PATCH v3 01/55] BZ#678544: audit: __link_path_walk should call audit_inode with original "name"
Bugzilla: 678544
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Currently, it modifies the "name" pointer in the process of doing its
normal work, and then calls audit_inode with the modified name. It
should instead be calling it with the original "name" that it was
passed.

This patch is not upstream. The code was substantially reworked there.

Signed-off-by: Jeff Layton <jlayton@redhat.com>