audit: optimize audit_compare_dname_path
Patchwork-id: 56959
O-Subject: [RHEL6.5 PATCH v3 10/55] BZ#678544: audit: optimize audit_compare_dname_path
Bugzilla: 678544
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

In the cases where we already know the length of the parent, pass it as
a parm so we don't need to recompute it. In the cases where we don't
know the length, pass in AUDIT_NAME_FULL (-1) to indicate that it should
be determined.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Upstream commit: e3d6b07b8ba161f638b026feba0c3c97875d7f1c