audit: overhaul __audit_inode_child to accomodate retrying
Patchwork-id: 56931
O-Subject: [RHEL6.5 PATCH v3 11/55] BZ#678544: audit: overhaul __audit_inode_child to accomodate retrying
Bugzilla: 678544
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

In order to accomodate retrying path-based syscalls, we need to add a
new "type" argument to audit_inode_child. This will tell us whether
we're looking for a child entry that represents a create or a delete.

If we find a parent, don't automatically assume that we need to create a
new entry. Instead, use the information we have to try to find an
existing entry first. Update it if one is found and create a new one if
not.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Upstream commit: 4fa6b5ecbf092c6ee752ece8a55d71f663d23254