vfs: add a retry_estale helper function to handle retries on ESTALE
Patchwork-id: 56946
O-Subject: [RHEL6.5 PATCH v3 26/55] BZ#678544: vfs: add a retry_estale helper function to handle retries on ESTALE
Bugzilla: 678544
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

This function is expected to be called from path-based syscalls to help
them decide whether to try the lookup and call again in the event that
they got an -ESTALE return back on an earier try.

Currently, we only retry the call once on an ESTALE error, but in the
event that we decide that that's not enough in the future, we should be
able to change the logic in this helper without too much effort.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>
Upstream commit: b9d6ba94b875192ef5e2dab92d72beea33b83c3d