percpu-rw-sem: remove the percpu rw semaphore implementation
Patchwork-id: 57088
O-Subject: [RHEL6.5 PATCH 19/19] kernel: remove the percpu rw semaphore implementation
Bugzilla: 909482
RH-Acked-by: Zach Brown <zab@redhat.com>

The percpu rwsem was added to protect access to the block device's
block_size member.  The previous patches in this patch series have
removed the reliance on locking for access to this structure member.
Since the only user of the code has been removed, let's get rid of
the semaphore implementation, too.

This is related to bug 909482.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jarod Wilson <jarod@redhat.com>