fuse: set page_descs length in fuse_buffered_write()
Patchwork-id: 57116
O-Subject: [RHEL6.5 PATCH] fuse: set page_descs length in fuse_buffered_write()
Bugzilla: 915135
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>
RH-Acked-by: Frantisek Hrbata <fhrbata@redhat.com>

The following rhel backport introduced the use of the page_descs
length field to describe the size of a fuse request:

	[fs] fuse: use req->page_descs[] for argpages cases

... but failed to set the field field in the fuse_write_end() (to
fuse_buffered_write()) codepath. This is a backport error as this
codepath does not exist in upstream fuse. The result of the error is
that requests submitted via this codepath hit fuse_copy_pages()
with a length of 0 and issue requests to userspace with a
zero-filled page. The fix is to set the length prior to submitting
the request.

Signed-off-by: Brian Foster <bfoster@redhat.com>