nfs: Ensure that we check lock exclusive/shared type against open modes
Patchwork-id: 57154
O-Subject: [RHEL6.5 PATCH BZ 916324] NFSv4: Ensure that we check lock exclusive/shared type against open modes
Bugzilla: 916324
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=916324

Fix NFSv4 regression seen by a lot of people so far on RHEL6.4 and the
latest RHEL6.3.z.  Regression snuck in by a small part of
commit aaccb96 which went in to fix bz 842435.

Patch is already upstream (commit 5572551) and clean backport.

This should go to 6.4.z as well as 6.3.z as well.

Testcase on the bugzilla and very easy to reproduce.
Mount a remote NFS filesystem using NFS4, then get a flock:
1.  mkdir /mnt/mountpoint
2.  mount -t nfs4 nfsserver:/dir /mnt/mountpoint
3.  cd /mnt/mountpoint
4.  strace flock -nx test.lock -c date
The command never returns
tail /var/log/messages to see stream of kernel messages

What we're doing above is opening the file with read permissions,
and trying an exclusive flock on it.  Apparently this is fairly
common for certain applications to do this.  On NFSv4, this operation
fails because we use byte range locks to simulate flock.
Before the offending commit, we let the operation go through down
to sending the RPC, the RPC came back with NFS4ERR_OPENMODE which
eventually got mapped to -EIO and exited cleanly.
After the commit, we go into an infinite loop retrying the RPC
and getting the same error back.

The right thing to do is to check for mismatch between the FMODE
and the type of the lock and if there's a mismatch, just return an
EBADF, before even trying to send out the RPC.  This is what upstream
does for this case.

I verified the patch fixes the issue myself with the testcase
and building a test kernel with the regressed kernels plus the
patch.  The kernels I tested the patch on were 2.6.32-358.2.1.el6
and 2.6.32-279.22.1.el6.  Patch also verified by one customer
who added the patch to 2.6.32-358.0.1.el6.

commit 55725513b5ef9d462aa3e18527658a0362aaae83
Author: Trond Myklebust <Trond.Myklebust@netapp.com>
Date:   Wed Apr 18 12:48:35 2012 -0400
Subject: [PATCH] NFSv4: Ensure that we check lock exclusive/shared type against open modes

Since we may be simulating flock() locks using NFS byte range locks,
we can't rely on the VFS having checked the file open mode for us.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Cc: stable@vger.kernel.org
Signed-off-by: Jarod Wilson <jarod@redhat.com>