defer do_filp_open() access checks to may_open()
Patchwork-id: 57491
O-Subject: [PATCH RHEL6.5] fs: defer do_filp_open() access checks to may_open()
Bugzilla: 920752
RH-Acked-by: Carlos Maiolino <cmaiolino@redhat.com>

This is for bug 920752 - cannot open device nodes for writing on RO filesystems

The big fsfreeze patchset in RHEL6.4, which contained

    1d3c71f [fs] Push mnt_want_write() outside of i_mutex

caused do_filp_open() to exit with an error early, if any write access
was requested on a readonly filesystem.  This is fine most of the time,
but not if we are trying to open a device node on the filesystem, for
example.  That should succeed.

echo foo > /mnt/test/devnull where devnull is the null device
fails, because it's opened O_CREAT, and that hits the failing
mnt_want_write (even though in the end the node does not need
to be created).

So defer the failures from mnt_want_write() until we potentially
get through i.e. may_open(), to see if such an action is allowed
after all.

Patch & analysis from Mateusz Guzik <mguzik@redhat.com>, thanks!

I've written an xfstest for this case and sent it upstream.

I've also run the change through xfstests and the pjd-fstest
posix suite with no regressions.

This is not strictly an upstream patch, but it follows a similar
mechanism to what is upstream.

I'd appreciate a couple careful reviews though, in any case.

Thanks,
-Eric

Signed-off-by: Eric Sandeen <sandeen@redhat.com>