irq: add quirk for broken interrupt remapping on 55XX chipsets
Patchwork-id: 57621
O-Subject: [RHEL 6.5 PATCH] irq: add quirk for broken interrupt remapping on 55XX chipsets [bz 887006]
Bugzilla: 887006
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

A few years back intel published a spec update:
http://www.intel.com/content/dam/doc/specification-update/5520-and-5500-chipset-ioh-specification-update.pdf

For the 5520 and 5500 chipsets which contained an errata (specificially errata
53), which noted that these chipsets can't properly do interrupt remapping, and
as a result the recommend that interrupt remapping be disabled in bios.  While
many vendors have a bios update to do exactly that, not all do, and of course
not all users update their bios to a level that corrects the problem.  As a
result, occasionally interrupts can arrive at a cpu even after affinity for that
interrupt has be moved, leading to lost or spurrious interrupts (usually
characterized by the message:
kernel: do_IRQ: 7.71 No irq handler for vector (irq -1)

There have been several incidents recently of people seeing this error, and
investigation has shown that they have system for which their BIOS level is such
that this feature was not properly turned off.  As such, it would be good to
give them a reminder that their systems are vulnurable to this problem.

BZ: 887006

Upstream: Intel has given its ok, but not outright ACK-ed it.  We're pushing
intel to send official ACK's but they seem to not want to touch this too much.
The Dons (Dutile and Zickus) are working on a possible workaround with Intel,
but given their stance seems to be a firm recommendation that remapping be
disabled, I think we need to take this as a RHEL6 only patch sooner rather than
later.

Brew: http://brewweb.devel.redhat.com/brew/taskinfo?taskID=5538059
Signed-off-by: Jarod Wilson <jarod@redhat.com>