gfs2: Allocate reservation structure before rename and link
Patchwork-id: 57715
O-Subject: [RHEL6.5 PATCH] [bz#922999] GFS2: Allocate reservation structure before rename and link
Bugzilla: 922999
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

Hi,

This patch allocates block reservation structures during rename and
link inode operations. Without this patch, GFS2 is susceptible to
a segfault / kernel panic if renames or symlinks are done without
prior allocations to the directory in which the ops are done.

The added lines of code already exist in the upstream kernel, but were
missing from RHEL6. The deleted lines of code were redundant and not
in upstream, so really this patch is just syncing a bit of code in RHEL6
with what's in the upstream GFS2.

It was tested on system gfs-a16c-02.mpc.lab.eng.bos.redhat.com. A test
kernel with this patch was also given to the customer who discovered the
problem. Results from their tests are still pending.

rhbz#922999

Regards,

Bob Peterson
Red Hat File Systems

Signed-off-by: Bob Peterson <rpeterso@redhat.com>