io_apic: initialize nr_ioapic_registers early in mp_register_ioapic()
Patchwork-id: 57732
O-Subject: [RHEL6.5 PATCH v3 1/1] x86, ioapic: initialize nr_ioapic_registers early in mp_register_ioapic()
Bugzilla: 855057
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=855057
Bug 855057 - RHEL6 can't boot installer with x2apic enabled in HWM6.4/6.6

The reason the entire upstream patch was not included is that the
number of io_apics is calculated and stored differently upstream,
and that part of the patch would break RHEL6.

Some of the arithmetic in mp_register_ioapic() is offset by one
from the upstream code because of the difference in the way the
gsi_base argument is calculated upstream.

Upstream discussion: https://lkml.org/lkml/2012/10/24/473

Upstream status:

commit 7716a5c4ff5f1f3dc5e9edcab125cbf7fceef0af
Author: Eric W. Biederman <ebiederm@xmission.com>
Date:   Tue Mar 30 01:07:12 2010 -0700

    x86, ioapic: Move nr_ioapic_registers calculation to mp_register_ioapic.

    Now that all ioapic registration happens in mp_register_ioapic we can
    move the calculation of nr_ioapic_registers there from enable_IO_APIC.
    The number of ioapic registers is already calucated in mp_register_ioapic
    so all that really needs to be done is to save the caluclated value
    in nr_ioapic_registers.

    Signed-off-by: Eric W. Biederman <ebiederm@xmission.com>
    LKML-Reference: <1269936436-7039-11-git-send-email-ebiederm@xmission.com>
    Signed-off-by: H. Peter Anvin <hpa@zytor.com>

Tested-by: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Tony Camuso <tcamuso@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>