perf, arch: Rework perf_event_index()
Patchwork-id: 57801
O-Subject: [PATCH RHEL6.5 BZ927177 19/22] perf, arch: Rework perf_event_index()
Bugzilla: 927177
RH-Acked-by: Myron Stowe <mstowe@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 927177
https://bugzilla.redhat.com/show_bug.cgi?id=927177

upstream status
===============
commit 35edc2a5095efb189e60dc32bbb9d2663aec6d24
Author: Peter Zijlstra <a.p.zijlstra@chello.nl>
Date:   Sun Nov 20 20:36:02 2011 +0100

description
===========
Put the logic to compute the event index into a per pmu method. This
is required because the x86 rules are weird and wonderful and don't
match the capabilities of the current scheme.

AFAIK only powerpc actually has a usable userspace read of the PMCs
but I'm not at all sure anybody actually used that.

ARM is restored to the default since it currently does not support
userspace access at all. And all software events are provided with a
method that reports their index as 0 (disabled).

Signed-off-by: Jarod Wilson <jarod@redhat.com>