perf/header: Set the tracepoint names on PERF_RECORD_HEADER_TRACING_DATA
Patchwork-id: 57840
O-Subject: [PATCH RHEL6.5 BZ927866 004/264] perf header: Set the tracepoint names on PERF_RECORD_HEADER_TRACING_DATA
Bugzilla: 927866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 927866
https://bugzilla.redhat.com/show_bug.cgi?id=927866

upstream status
===============
commit 8b6ee4c5d48d93527dcf6e36c51cbb7703d7fffb
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Tue Aug 7 23:36:16 2012 -0300

description
===========
We only have access to pevent after processing that event, so set the
tracepoint names there.

Right now this isn't a problem as we're deferring resolving the
tracepoint names to when we process samples, but in the next patches we
will be doing it in advance, to avoid relookups, so do it earlier, as
soon as we process the tracing data event.

Signed-off-by: Jarod Wilson <jarod@redhat.com>