perf/test: Add round trip test for sw and hw event names
Patchwork-id: 57920
O-Subject: [PATCH RHEL6.5 BZ927866 103/264] perf test: Add round trip test for sw and hw event names
Bugzilla: 927866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 927866
https://bugzilla.redhat.com/show_bug.cgi?id=927866

upstream status
===============
commit 8ad7013b252ba683055df19e657eb03d98f4f312
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Thu Sep 6 13:11:18 2012 -0300

description
===========
It basically traverses the hardware and software event name arrays
creating an evlist with all events, then it uses perf_evsel__name to
check that the name is the expected one.

With it I noticed this problem:

[root@sandy ~]# perf test 10
10: roundtrip evsel->name check:invalid or unsupported event: 'CPU-migrations'
Run 'perf list' for a list of valid events
 FAILED!

Changed it to "cpu-migrations" in the software event arrays and it
worked.

This is to catch problems like the one reported by Joel Uckelman in
http://permalink.gmane.org/gmane.linux.kernel.perf.user/1016

Hardware cache events will be checked in the following patch.