perf: Use __maybe_used for unused variables
Patchwork-id: 57978
O-Subject: [PATCH RHEL6.5 BZ927866 141/264] perf tools: Use __maybe_used for unused variables
Bugzilla: 927866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Irina Tirdea <irina.tirdea@gmail.com>

Bugzilla: 927866
https://bugzilla.redhat.com/show_bug.cgi?id=927866

upstream status
===============
commit 1d037ca1648b775277fc96401ec2aa233724906c
Author: Irina Tirdea <irina.tirdea@gmail.com>
Date:   Tue Sep 11 01:15:03 2012 +0300

description
===========
perf defines both __used and __unused variables to use for marking
unused variables. The variable __used is defined to
__attribute__((__unused__)), which contradicts the kernel definition to
__attribute__((__used__)) for new gcc versions. On Android, __used is
also defined in system headers and this leads to warnings like: warning:
'__used__' attribute ignored

__unused is not defined in the kernel and is not a standard definition.
If __unused is included everywhere instead of __used, this leads to
conflicts with glibc headers, since glibc has a variables with this name
in its headers.

The best approach is to use __maybe_unused, the definition used in the
kernel for __attribute__((unused)). In this way there is only one
definition in perf sources (instead of 2 definitions that point to the
same thing: __used and __unused) and it works on both Linux and Android.
This patch simply replaces all instances of __used and __unused with
__maybe_unused.

[ committer note: fixed up conflict with a116e05 in builtin-sched.c ]