perf: remove sscanf extension %as
Patchwork-id: 57997
O-Subject: [PATCH RHEL6.5 BZ927866 173/264] perf tools: remove sscanf extension %as
Bugzilla: 927866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Irina Tirdea <irina.tirdea@intel.com>

Bugzilla: 927866
https://bugzilla.redhat.com/show_bug.cgi?id=927866

upstream status
===============
commit bcbd004020bf0d725722be75da35fd326ff63ef4
Author: Irina Tirdea <irina.tirdea@intel.com>
Date:   Thu Sep 20 23:37:50 2012 +0300

description
===========
perf uses sscanf extension %as to read and allocate a string in the same
step.  This is a non-standard extension only present in new versions of
glibc.

Replacing the use of sscanf and %as with strtok_r calls in order to
parse a given string into its components.  This is needed in Android
since bionic does not support
%as extension for sscanf.