perf/inject: Remove static variables
Patchwork-id: 58024
O-Subject: [PATCH RHEL6.5 BZ927866 215/264] perf inject: Remove static variables
Bugzilla: 927866
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Ivan Vecera <ivecera@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Arnaldo Carvalho de Melo <acme@redhat.com>

Bugzilla: 927866
https://bugzilla.redhat.com/show_bug.cgi?id=927866

upstream status
===============
commit 5ded57ac1bdb99b716c1da96eb8b5d387f5eb676
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date:   Sun Sep 30 19:54:10 2012 -0300

description
===========
We want to reduce the impact that each of the builtins has on perf as a
whole, so use the superclassing of perf_tool mechanizm to move its
config knobs to the stack, so that only if we use that tool, its impact
will be felt.

In this case is more about consistency, as the impact of this tool is
minimal.